Merge pull request !1 from zeki_zhao/develop

+++ arch/arm/src/lpc31xx/lpc31_timerisr.c
@@ -97,7 +97,7 @@ void up_timer_initialize(void)
    * of the timer0 module clock (in the AHB0APB1_BASE domain (2)).
    */
 
-  freq = (uint64_t)lpc31_clkfreq(CLKID_TIMER0PCLK, DOMAINID_AHB0APB1);
+  freq = lpc31_clkfreq(CLKID_TIMER0PCLK, DOMAINID_AHB0APB1);
 
   /* If the clock is >1MHz, use pre-dividers */
 
@@ -111,7 +111,7 @@ void up_timer_initialize(void)
       freq   >>= 4;
     }
 
-  load = ((freq * (uint64_t)10000) / 1000000);
+  load = ((freq * 10000) / 1000000);
   putreg32((uint32_t)load, LPC31_TIMER0_LOAD);
 
   /* Set periodic mode */
